pyproject.toml
[project]
name = "ai-event-finder-benchmark"
version = "0.0.1"
description = "Evaluate various AI models based on their ability to identify events."
readme = "README.md"
requires-python = ">=3.11.2"
dependencies = [
    "anthropic==0.60.0",
    "fuzzywuzzy==0.18.0",
    "geopy==2.4.1",
    "google-genai==1.28.0",
    "matplotlib==3.10.5",
    "openai==1.98.0",
    "pandas==2.3.1",
    "python-dotenv==1.0.0",
    "python-levenshtein==0.27.1",
    "seaborn==0.13.2",
    "streamlit==1.48.0",
]
